Reuse getPageDevice method for MetadataFile and MetadataDummy
This fixes an issue where a bug had been fixed only in Dummy and not in File.
